+++ app/src/main/java/com/example/socialcompass/IOrientation.java
@@ -5,5 +5,5 @@ public interface IOrientation {
     // where 0 is north, pi/2 is east, -pi/2 is west, etc.
     // ideally, this would return a value in the range [-pi, pi]
     // which is the range of android's sensormanager.getorientation()
-    double getOrientation();
+    float getOrientation();
 }
